requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# LLM interface
litellm>=1.30.0

# Analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Utilities
pyyaml>=6.0
jsonlines>=3.1.0